requirements.txt
numpy==1.25.2
Pillow==10.0.0
tk==0.1.0
torch==2.0.1
torchvision==0.15.2
typing_extensions==4.7.1